重构 vote 模型代码
